no more references to Signer/MessageSigner in proto
