+++ appveyor.yml
@@ -33,9 +33,9 @@ test_script:
   - cargo test --manifest-path client/Cargo.toml --all-features
   - cargo test --manifest-path client/Cargo.toml --no-default-features
   - cargo test --manifest-path client/Cargo.toml
-  - cargo test --manifest-path client/Cargo.toml --no-default-features --features=openssl
+  - cargo test --manifest-path client/Cargo.toml --no-default-features --features=dnssec-openssl
   - cargo test --manifest-path client/Cargo.toml --no-default-features
-  - cargo test --manifest-path client/Cargo.toml --no-default-features --features=ring
+  - cargo test --manifest-path client/Cargo.toml --no-default-features --features=dnssec-ring
 
   - cargo test --manifest-path rustls/Cargo.toml
   - cargo test --manifest-path openssl/Cargo.toml
