+++ proto/src/op/mod.rs
@@ -27,7 +27,7 @@ pub mod response_code;
 pub use self::edns::Edns;
 pub use self::header::Header;
 pub use self::header::MessageType;
-pub use self::message::{Message, UpdateMessage};
+pub use self::message::{Message, MessageFinalizer};
 pub use self::op_code::OpCode;
 pub use self::query::Query;
 pub use self::response_code::ResponseCode;
