+++ proto/src/rr/dnssec/tbs.rs
@@ -297,106 +297,3 @@ pub fn determine_name(name: &Name, num_labels: u8) -> Option<Name> {
     // TODO: this should be an error
     None
 }
-
-#[cfg(feature = "openssl")]
-#[cfg(test)]
-mod tests {
-    extern crate openssl;
-    use self::openssl::rsa::Rsa;
-
-    use rr::{Name, RecordType};
-    use rr::rdata::SIG;
-    use rr::dnssec::{KeyPair, Signer};
-
-    pub use super::*;
-
-    #[test]
-    fn test_rrset_tbs() {
-        let rsa = Rsa::generate(2048).unwrap();
-        let key = KeyPair::from_rsa(rsa).unwrap();
-        let sig0key = key.to_sig0key(Algorithm::RSASHA256).unwrap();
-        let signer = Signer::sig0(sig0key, key, Name::root());
-
-        let origin: Name = Name::parse("example.com.", None).unwrap();
-        let rrsig = Record::new()
-            .set_name(origin.clone())
-            .set_ttl(86400)
-            .set_rr_type(RecordType::NS)
-            .set_dns_class(DNSClass::IN)
-            .set_rdata(RData::SIG(SIG::new(
-                RecordType::NS,
-                Algorithm::RSASHA256,
-                origin.num_labels(),
-                86400,
-                5,
-                0,
-                signer.calculate_key_tag().unwrap(),
-                origin.clone(),
-                vec![],
-            )))
-            .clone();
-        let rrset = vec![
-            Record::new()
-                .set_name(origin.clone())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::NS)
-                .set_dns_class(DNSClass::IN)
-                .set_rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()))
-                .clone(),
-            Record::new()
-                .set_name(origin.clone())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::NS)
-                .set_dns_class(DNSClass::IN)
-                .set_rdata(RData::NS(Name::parse("b.iana-servers.net.", None).unwrap()))
-                .clone(),
-        ];
-
-        let tbs = rrset_tbs_with_rrsig(&rrsig, &rrset).unwrap();
-        assert!(!tbs.0.is_empty());
-
-        let rrset = vec![
-            Record::new()
-                .set_name(origin.clone())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::CNAME)
-                .set_dns_class(DNSClass::IN)
-                .set_rdata(RData::CNAME(
-                    Name::parse("a.iana-servers.net.", None).unwrap(),
-                ))
-                .clone(), // different type
-            Record::new()
-                .set_name(Name::parse("www.example.com.", None).unwrap())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::NS)
-                .set_dns_class(DNSClass::IN)
-                .set_rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()))
-                .clone(), // different name
-            Record::new()
-                .set_name(origin.clone())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::NS)
-                .set_dns_class(DNSClass::CH)
-                .set_rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()))
-                .clone(), // different class
-            Record::new()
-                .set_name(origin.clone())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::NS)
-                .set_dns_class(DNSClass::IN)
-                .set_rdata(RData::NS(Name::parse("a.iana-servers.net.", None).unwrap()))
-                .clone(),
-            Record::new()
-                .set_name(origin.clone())
-                .set_ttl(86400)
-                .set_rr_type(RecordType::NS)
-                .set_dns_class(DNSClass::IN)
-                .set_rdata(RData::NS(Name::parse("b.iana-servers.net.", None).unwrap()))
-                .clone(),
-        ];
-
-        let filtered_tbs = rrset_tbs_with_rrsig(&rrsig, &rrset).unwrap();
-        assert!(!filtered_tbs.0.is_empty());
-        assert_eq!(tbs.as_ref(), filtered_tbs.as_ref());
-    }
-}
